Devide chrono block and absctract block
